+++ src/components/MyPage_Component/ProfileCard.jsx
@@ -1,7 +1,10 @@
 import React from 'react'
 import ProfileBasic from '../../images/ProfileBasic.png'
-import * as T from '../MyPage'
+import * as T from './MyPage'
 import styled from 'styled-components'
+import Header from '../header/index'
+import Footer from '../footer/index'
+import Title from '../title/index'
 
 const NameH4 = styled.h4`
     font-size: 36px;
@@ -63,31 +66,44 @@ const SelfIntroductionP = styled.p`
   font-size: 20px;
 `;
 
+const FooterContainer = styled.div`
+  position: absolute;
+  bottom: -1200px;
+  width: 100%;
+`;
+
 export default function Profile_card() {
   return (
-    // 하드 코딩
-    <T.TotalDiv>
-      <T.GlobalStyle/>
-      <T.TitleH3>제니 님의 프로필 카드</T.TitleH3>
-      <SectionContainer>
-          <SectionDiv>
-            <ProfileBasicImg src={ProfileBasic} alt ="ProfileBasic"/>
-            <NameH4>제니</NameH4>
-            <InfoP>년생 : 00년생</InfoP>
-            <InfoP>키 : 163cm</InfoP>
-            <InfoP>지역 : 부산광역시 남구</InfoP>
-          </SectionDiv>
-          <SectionDiv>
-            <InfoP>장거리 가능 여부 : 불가능</InfoP>
-            <InfoP>흡연 여부 : 비흡연</InfoP>
-            <InfoP>음주 여부 : 음주</InfoP>
-            <InfoP>단과대 : 경영대학</InfoP>
-              <SelfIntroductionTitleP>자기 소개</SelfIntroductionTitleP>
-              <SelfIntroductionDiv>
-                <SelfIntroductionP> 처음에는 많이 뚝딱거릴 수도 있지만 친해지면 엄청 애교도 많고 활발해요! 좋은 인연 만들어 가고 싶어요!</SelfIntroductionP>       
-            </SelfIntroductionDiv>
-          </SectionDiv>
-      </SectionContainer>
-    </T.TotalDiv>
+    <>
+      <Header/>
+      <Title title="마이페이지"/>
+      <T.TotalHr></T.TotalHr>
+      <T.TotalDiv>
+        <T.TitleH3>제니 님의 프로필 카드</T.TitleH3>
+        <SectionContainer>
+            <SectionDiv>
+              <ProfileBasicImg src={ProfileBasic} alt ="ProfileBasic"/>
+              <NameH4>제니</NameH4>
+              <InfoP>년생 : 00년생</InfoP>
+              <InfoP>키 : 163cm</InfoP>
+              <InfoP>지역 : 부산광역시 남구</InfoP>
+            </SectionDiv>
+            <SectionDiv>
+              <InfoP>장거리 가능 여부 : 불가능</InfoP>
+              <InfoP>흡연 여부 : 비흡연</InfoP>
+              <InfoP>음주 여부 : 음주</InfoP>
+              <InfoP>단과대 : 경영대학</InfoP>
+                <SelfIntroductionTitleP>자기 소개</SelfIntroductionTitleP>
+                <SelfIntroductionDiv>
+                  <SelfIntroductionP> 처음에는 많이 뚝딱거릴 수도 있지만 친해지면 엄청 애교도 많고 활발해요! 좋은 인연 만들어 가고 싶어요!</SelfIntroductionP>       
+              </SelfIntroductionDiv>
+            </SectionDiv>
+        </SectionContainer>
+      </T.TotalDiv>
+      <FooterContainer>
+        <Footer/>
+      </FooterContainer>
+    </>
+    
   )
 }
